requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
playwright==1.58.0
python-dotenv==1.2.1